test(select): Remove unnecessary snapshots
